Add custom split view controller to be shared within the app
