+++ WooCommerce/Classes/System/WooSplitViewController.swift
@@ -1,0 +1,35 @@
+import UIKit
+
+/// Custom split view controller with double column style with preferred tile split behavior and 2-column display mode.
+/// When collapsed, the split view falls back to display the primary column.
+///
+final class WooSplitViewController: UISplitViewController {
+
+    init() {
+        super.init(style: .doubleColumn)
+        configureCommonStyle()
+    }
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    private func configureCommonStyle() {
+        preferredDisplayMode = .oneBesideSecondary
+        preferredSplitBehavior = .tile
+        delegate = self
+    }
+}
+
+extension WooSplitViewController: UISplitViewControllerDelegate {
+    func splitViewController(_ splitViewController: UISplitViewController,
+                             topColumnForCollapsingToProposedTopColumn proposedTopColumn: UISplitViewController.Column) -> UISplitViewController.Column {
+        // Always fall back to display the primary column when collapsed.
+        return .primary
+    }
+
+    func splitViewController(_ splitViewController: UISplitViewController, willChangeTo displayMode: UISplitViewController.DisplayMode) {
+        // Automatically hides the default toggle button if displaying 2 columns.
+        splitViewController.presentsWithGesture = displayMode != .oneBesideSecondary
+    }
+}
